Put Volta at the head of the path
Some systems come installed with a nodejs toolchain already in
`/usr/bin` or `/usr/local/bin`. The entire point of Volta is to
supercede those toolchains and so by necessity it must take
precedence.

This just moves it up to the head of the path so that volta toolchains
will mask any other toolchain that is in the default path.